Agregué barras de progreso y reorganicé archivos
