+++ src/components/PlanDeEstudio/BarrasDeProgreso/BarrasDeProgreso.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+export default function BarrasDeProgreso() {
+  return (
+    <div>
+      
+    </div>
+  )
+}
